requirements.txt
aiohttp>=3.8.0
python-dotenv>=0.19.0
openai>=1.0.0
google-generativeai>=0.3.0
python-telegram-bot>=20.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0